+++ image_proc/include/image_proc/crop_non_zero.hpp
@@ -1,0 +1,71 @@
+// Copyright 2008, 2019 Willow Garage, Inc., Steve Macenski, Joshua Whitley
+// All rights reserved.
+//
+// Software License Agreement (BSD License 2.0)
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions
+// are met:
+//
+// * Redistributions of source code must retain the above copyright
+//   notice, this list of conditions and the following disclaimer.
+// * Redistributions in binary form must reproduce the above
+//   copyright notice, this list of conditions and the following
+//   disclaimer in the documentation and/or other materials provided
+//   with the distribution.
+// * Neither the name of {copyright_holder} nor the names of its
+//   contributors may be used to endorse or promote products derived
+//   from this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+// FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+// COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+// INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+// BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+// LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+// ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
+#ifndef IMAGE_PROC__CROP_NON_ZERO_HPP_
+#define IMAGE_PROC__CROP_NON_ZERO_HPP_
+
+#include <cv_bridge/cv_bridge.h>
+#include <image_transport/image_transport.h>
+#include <rclcpp/rclcpp.hpp>
+
+#include <thread>
+#include <memory>
+#include <vector>
+#include <string>
+
+namespace image_proc
+{
+
+class CropNonZeroNode : public rclcpp::Node
+{
+public:
+  explicit CropNonZeroNode(const rclcpp::NodeOptions &);
+
+private:
+  std::string camera_namespace_;
+  std::string image_pub_topic_;
+  std::string image_sub_topic_;
+
+  // Subscriptions
+  image_transport::Subscriber sub_raw_;
+
+  // Publications
+  std::mutex connect_mutex_;
+
+  image_transport::Publisher pub_;
+
+  void connectCb();
+
+  void imageCb(const sensor_msgs::msg::Image::ConstSharedPtr & raw_msg);
+};
+}  // namespace image_proc
+#endif  // IMAGE_PROC__CROP_NON_ZERO_HPP_
